+++ guestbook/.idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/helloworld.iml" filepath="$PROJECT_DIR$/.idea/helloworld.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/guestbook.iml" filepath="$PROJECT_DIR$/.idea/guestbook.iml" />
     </modules>
   </component>
 </project>
