--- helloworld/main.py
@@ -1,21 +1,0 @@
-from google.appengine.api import users
-
-import webapp2
-
-
-class MainPage(webapp2.RequestHandler):
-
-    def get(self):
-        # Checks for active Google account session
-        user = users.get_current_user()
-
-        if user:
-            self.response.headers['Content-Type'] = 'text/html; charset=utf-8'
-            self.response.write('Hello, ' + user.nickname())
-        else:
-            self.redirect(users.create_login_url(self.request.uri))
-
-
-app = webapp2.WSGIApplication([
-    ('/', MainPage),
-], debug=True)
